Testing a fix for 64-bit pointer arithmetic.

+++ src-library/ScorePageBaseSimple.cpp
@@ -166,7 +166,7 @@ int ScorePageBaseSimple::findStaff(int staffno) {
    if (searchresult == NULL) {
       return -1;
    } else {
-      int start = ((int)searchresult - (int)data.getBase())/sizeof(ScoreRecord);
+      int start = (int)((searchresult - data.getBase())/sizeof(ScoreRecord));
       int i;
       for (i=start; i>=0; i++) {
          if ((int)data[i].getPValue(2) != staffno) {
